backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
chromadb
sentence-transformers
huggingface_hub
python-dotenv
requests
